Issue #380 clustering evac and room exit weights bug (#397)
* clustering evac and room exit weights bug but also new mechanism for navmesh smoke avoidance and other fixes

* small fixes

* small fix

* restore pepper js

+++ installer/sql.sql
@@ -43,7 +43,7 @@ CREATE TABLE simulations ( ---{{{
     fireorigname text,
     heat_detectors text,
     smoke_detectors text,
-    hrrpeak smallint,
+    hrrpeak int,
     soot_yield decimal,
     co_yield decimal,
     hcl_yield decimal,
